refactor: moved componets and their module.css files into individual component folders. Create basic file structure.
