--- src/TodoList.module.css
@@ -1,39 +1,0 @@
-.TodoList {
-  background-color: var(--light-background);
-  border-radius: 1rem;
-  max-width: 25rem;
-  margin: auto;
-  padding: 2rem 1rem;
-  text-align: center;
-}
-
-input[type='checkbox' i]:checked + label {
-  text-decoration: line-through;
-}
-
-ul {
-  list-style: none;
-  align-content: right;
-}
-
-.btn {
-  appearance: none;
-  font-family: inherit;
-  font-size: 100%;
-  padding: 0.3em;
-  margin-left: auto;
-  border-radius: 4px;
-  border-style: solid;
-  border-width: 0.1rem;
-}
-
-.remove {
-  background-color: var(--light-background);
-  color: var(--neutral-blue);
-  border-color: var(--neutral-blue);
-}
-
-.remove:hover {
-  color: var(--negative-red);
-  border-color: var(--negative-red);
-}
