+++ src/components/AddTodoForm/AddTodoForm.module.css
@@ -26,14 +26,3 @@ input[type='text' i] {
   border-style: solid;
   border-width: 0.1rem;
 }
-
-.btn {
-  appearance: none;
-  font-family: inherit;
-  font-size: 100%;
-  padding: 0.3em;
-  margin-left: auto;
-  border-radius: 4px;
-  border-style: solid;
-  border-width: 0.1rem;
-}
