+++ src/components/TodoList/TodoList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import TodoListItem from './TodoListItem';
+import TodoListItem from '../TodoListItem/TodoListItem';
 import styles from './TodoList.module.css';
 
 function TodoList({ todoList, onRemoveTodo }) {
